Move k8s integration related code to k8s/clusterExplorer.ts
